--- a/src/FirebirdSql.Data.FirebirdClient/FirebirdClient/FbBatchCommand.cs
+++ b/src/FirebirdSql.Data.FirebirdClient/FirebirdClient/FbBatchCommand.cs
@@ -19,12 +19,11 @@ using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data;
-using System.Data.Common;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using FirebirdSql.Data.Common;
 using FirebirdSql.Data.Logging;
+using Microsoft.Extensions.Logging;
 
 namespace FirebirdSql.Data.FirebirdClient;
 
@@ -33,7 +32,7 @@ public sealed class FbBatchCommand : IFbPreparedCommand, IDescriptorFiller, IDis
 		, IAsyncDisposable
 #endif
 {
-	static readonly IFbLogger Log = FbLogManager.CreateLogger(nameof(FbBatchCommand));
+	static readonly ILogger<FbBatchCommand> Log = FbLogManager.CreateLogger<FbBatchCommand>();
 
 	private const int DefaultBatchBufferSize = 16 * 1024 * 1024;