--- a/src/FirebirdSql.Data.FirebirdClient/FirebirdClient/FbConnection.cs
+++ b/src/FirebirdSql.Data.FirebirdClient/FirebirdClient/FbConnection.cs
@@ -23,13 +23,14 @@ using System.Threading;
 using System.Threading.Tasks;
 using FirebirdSql.Data.Common;
 using FirebirdSql.Data.Logging;
+using Microsoft.Extensions.Logging;
 
 namespace FirebirdSql.Data.FirebirdClient;
 
 [DefaultEvent("InfoMessage")]
 public sealed class FbConnection : DbConnection, ICloneable
 {
-	static readonly IFbLogger Log = FbLogManager.CreateLogger(nameof(FbConnection));
+	static readonly ILogger<FbConnection> Log = FbLogManager.CreateLogger<FbConnection>();
 
 	#region Static Pool Handling Methods