--- a/src/FirebirdSql.Data.FirebirdClient/Logging/FbLogManager.cs
+++ b/src/FirebirdSql.Data.FirebirdClient/Logging/FbLogManager.cs
@@ -15,37 +15,26 @@
 
 //$Authors = Jiri Cincura (jiri@cincura.net)
 
-using System;
+using Microsoft.Extensions.Logging.Abstractions;
+using Microsoft.Extensions.Logging;
 
 namespace FirebirdSql.Data.Logging;
 
 public static class FbLogManager
 {
-	public static IFbLoggingProvider Provider
-	{
-		get
-		{
-			_providerRetrieved = true;
-			return _provider;
-		}
-		set
-		{
-			if (_providerRetrieved)
-				throw new InvalidOperationException("The logging provider must be set before any action is taken");
+	internal static ILoggerFactory LoggerFactory = NullLoggerFactory.Instance;
+	internal static bool IsParameterLoggingEnabled = false;
 
-			_provider = value ?? throw new ArgumentNullException(nameof(value));
-		}
+	public static void UseLoggerFactory(ILoggerFactory loggerFactory)
+	{
+		LoggerFactory = loggerFactory;
 	}
 
-	public static bool IsParameterLoggingEnabled { get; set; }
-
-	static IFbLoggingProvider _provider;
-	static bool _providerRetrieved;
-
-	static FbLogManager()
+	public static void EnableParameterLogging(bool enable = true)
 	{
-		_provider = new NullLoggingProvider();
+		IsParameterLoggingEnabled = enable;
 	}
 
-	internal static IFbLogger CreateLogger(string name) => Provider.CreateLogger("FirebirdClient." + name);
+	internal static ILogger<T> CreateLogger<T>() =>
+		LoggerFactory.CreateLogger<T>();
 }