--- a/src/FirebirdSql.Data.FirebirdClient/Logging/LogMessages.cs
+++ b/src/FirebirdSql.Data.FirebirdClient/Logging/LogMessages.cs
@@ -16,274 +16,122 @@
 //$Authors = Jiri Cincura (jiri@cincura.net)
 
 using System;
-using System.Text;
+using System.Collections.Generic;
+using System.Data.Common;
+using System.Linq;
 using FirebirdSql.Data.FirebirdClient;
+using Microsoft.Extensions.Logging;
 
 namespace FirebirdSql.Data.Logging;
 
 static class LogMessages
 {
-	public static void CommandExecution(IFbLogger log, FbCommand command)
+	public static void CommandExecution(ILogger log, FbCommand command)
 	{
-		if (!log.IsEnabled(FbLogLevel.Debug))
-			return;
-
-		var sb = new StringBuilder();
-		sb.AppendLine("Command execution:");
-		sb.AppendLine(command.CommandText);
-		if (FbLogManager.IsParameterLoggingEnabled)
+		if (!log.IsEnabled(LogLevel.Debug))
 		{
-			sb.AppendLine("Parameters:");
-			if (!command.HasParameters)
-			{
-				sb.AppendLine("<no parameters>");
-			}
-			else
-			{
-				foreach (FbParameter parameter in command.Parameters)
-				{
-					var name = parameter.ParameterName;
-					var type = parameter.FbDbType;
-					var value = !IsNullParameterValue(parameter.InternalValue) ? parameter.InternalValue : "<null>";
-					sb.AppendLine($"Name:{name}\tType:{type}\tUsed Value:{value}");
-				}
-			}
-		}
-		log.Debug(sb.ToString());
-	}
-	public static void CommandExecution(IFbLogger log, FbBatchCommand command)
-	{
-		if (!log.IsEnabled(FbLogLevel.Debug))
 			return;
+		}
 
-		var sb = new StringBuilder();
-		sb.AppendLine("Command execution:");
-		sb.AppendLine(command.CommandText);
-		if (FbLogManager.IsParameterLoggingEnabled)
+		if (!FbLogManager.IsParameterLoggingEnabled || !command.HasParameters)
 		{
-			sb.AppendLine("Parameters:");
-			if (command.HasParameters)
-			{
-				sb.AppendLine("<no parameters>");
-			}
-			else
-			{
-				foreach (var batchParameter in command.BatchParameters)
-				{
-					foreach (FbParameter parameter in batchParameter)
-					{
-						var name = parameter.ParameterName;
-						var type = parameter.FbDbType;
-						var value = !IsNullParameterValue(parameter.InternalValue) ? parameter.InternalValue : "<null>";
-						sb.AppendLine($"Name:{name}\tType:{type}\tUsed Value:{value}");
-					}
-				}
-			}
+			log.LogDebug("Command execution: {command}", command.CommandText);
+			return;
 		}
-		log.Debug(sb.ToString());
+
+		var parameters = FbParameterCollectionToDictionary(command.Parameters);
+		log.LogDebug("Command execution: {command}, {parameters}", command.CommandText, parameters);
 	}
 
-	public static void ConnectionOpening(IFbLogger log, FbConnection connection)
+	public static void CommandExecution(ILogger log, FbBatchCommand command)
 	{
-		if (!log.IsEnabled(FbLogLevel.Debug))
+		if (!log.IsEnabled(LogLevel.Debug))
+		{
 			return;
+		}
 
-		var sb = new StringBuilder();
-		sb.AppendLine("Opening connection:");
-		sb.AppendLine($"Connection String: {connection.ConnectionString}");
-		log.Debug(sb.ToString());
-	}
-	public static void ConnectionOpened(IFbLogger log, FbConnection connection)
-	{
-		if (!log.IsEnabled(FbLogLevel.Debug))
+		if (!FbLogManager.IsParameterLoggingEnabled || !command.HasParameters)
+		{
+			log.LogDebug("Command execution: {command}", command.CommandText);
 			return;
+		}
 
-		var sb = new StringBuilder();
-		sb.AppendLine("Opened connection:");
-		sb.AppendLine($"Connection String: {connection.ConnectionString}");
-		log.Debug(sb.ToString());
-	}
-	public static void ConnectionClosing(IFbLogger log, FbConnection connection)
-	{
-		if (!log.IsEnabled(FbLogLevel.Debug))
-			return;
-
-		var sb = new StringBuilder();
-		sb.AppendLine("Closing connection:");
-		sb.AppendLine($"Connection String: {connection.ConnectionString}");
-		log.Debug(sb.ToString());
-	}
-	public static void ConnectionClosed(IFbLogger log, FbConnection connection)
-	{
-		if (!log.IsEnabled(FbLogLevel.Debug))
-			return;
-
-		var sb = new StringBuilder();
-		sb.AppendLine("Closed connection:");
-		sb.AppendLine($"Connection String: {connection.ConnectionString}");
-		log.Debug(sb.ToString());
+		var parameters = command.BatchParameters.SelectMany(FbParameterCollectionToDictionary);
+		log.LogDebug("Command execution: {command}, {parameters}", command.CommandText, parameters);
 	}
 
-	public static void TransactionBeginning(IFbLogger log, FbTransaction transaction)
-	{
-		if (!log.IsEnabled(FbLogLevel.Debug))
-			return;
+	public static void ConnectionOpening(ILogger log, FbConnection connection) =>
+		log.LogDebug("Opening connection: {connectionString}", connection.ConnectionString);
 
-		var sb = new StringBuilder();
-		sb.AppendLine("Beginning transaction:");
-		sb.AppendLine($"Isolation Level: {transaction.IsolationLevel}");
-		log.Debug(sb.ToString());
-	}
-	public static void TransactionBegan(IFbLogger log, FbTransaction transaction)
-	{
-		if (!log.IsEnabled(FbLogLevel.Debug))
-			return;
+	public static void ConnectionOpened(ILogger log, FbConnection connection) =>
+		log.LogDebug("Opened connection: {connectionString}", connection.ConnectionString);
 
-		var sb = new StringBuilder();
-		sb.AppendLine("Began transaction:");
-		sb.AppendLine($"Isolation Level: {transaction.IsolationLevel}");
-		log.Debug(sb.ToString());
-	}
-	public static void TransactionCommitting(IFbLogger log, FbTransaction transaction)
-	{
-		if (!log.IsEnabled(FbLogLevel.Debug))
-			return;
+	public static void ConnectionClosing(ILogger log, FbConnection connection) =>
+		log.LogDebug("Closing connection: {connectionString}", connection.ConnectionString);
 
-		var sb = new StringBuilder();
-		sb.AppendLine("Committing transaction:");
-		sb.AppendLine($"Isolation Level: {transaction.IsolationLevel}");
-		log.Debug(sb.ToString());
-	}
-	public static void TransactionCommitted(IFbLogger log, FbTransaction transaction)
-	{
-		if (!log.IsEnabled(FbLogLevel.Debug))
-			return;
+	public static void ConnectionClosed(ILogger log, FbConnection connection) =>
+		log.LogDebug("Closed connection: {connectionString}", connection.ConnectionString);
 
-		var sb = new StringBuilder();
-		sb.AppendLine("Committed transaction:");
-		sb.AppendLine($"Isolation Level: {transaction.IsolationLevel}");
-		log.Debug(sb.ToString());
-	}
-	public static void TransactionRollingBack(IFbLogger log, FbTransaction transaction)
-	{
-		if (!log.IsEnabled(FbLogLevel.Debug))
-			return;
+	public static void TransactionBeginning(ILogger log, FbTransaction transaction) =>
+		// TODO: Transaction Id?
+		log.LogDebug("Beginning transaction: {isolationLevel}", transaction.IsolationLevel);
 
-		var sb = new StringBuilder();
-		sb.AppendLine("Rolling back transaction:");
-		sb.AppendLine($"Isolation Level: {transaction.IsolationLevel}");
-		log.Debug(sb.ToString());
-	}
-	public static void TransactionRolledBack(IFbLogger log, FbTransaction transaction)
-	{
-		if (!log.IsEnabled(FbLogLevel.Debug))
-			return;
+	public static void TransactionBegan(ILogger log, FbTransaction transaction) =>
+		log.LogDebug("Began transaction: {isolationLevel}", transaction.IsolationLevel);
 
-		var sb = new StringBuilder();
-		sb.AppendLine("Rolled back transaction:");
-		sb.AppendLine($"Isolation Level: {transaction.IsolationLevel}");
-		log.Debug(sb.ToString());
-	}
-	public static void TransactionSaving(IFbLogger log, FbTransaction transaction)
-	{
-		if (!log.IsEnabled(FbLogLevel.Debug))
-			return;
+	public static void TransactionCommitting(ILogger log, FbTransaction transaction) =>
+		log.LogDebug("Committing transaction: {isolationLevel}", transaction.IsolationLevel);
 
-		var sb = new StringBuilder();
-		sb.AppendLine("Creating savepoint:");
-		sb.AppendLine($"Isolation Level: {transaction.IsolationLevel}");
-		log.Debug(sb.ToString());
-	}
-	public static void TransactionSaved(IFbLogger log, FbTransaction transaction)
-	{
-		if (!log.IsEnabled(FbLogLevel.Debug))
-			return;
+	public static void TransactionCommitted(ILogger log, FbTransaction transaction) =>
+		log.LogDebug("Committed transaction: {isolationLevel}", transaction.IsolationLevel);
 
-		var sb = new StringBuilder();
-		sb.AppendLine("Created savepoint:");
-		sb.AppendLine($"Isolation Level: {transaction.IsolationLevel}");
-		log.Debug(sb.ToString());
-	}
-	public static void TransactionReleasingSavepoint(IFbLogger log, FbTransaction transaction)
-	{
-		if (!log.IsEnabled(FbLogLevel.Debug))
-			return;
+	public static void TransactionRollingBack(ILogger log, FbTransaction transaction) =>
+		log.LogDebug("Rolling back transaction: {isolationLevel}", transaction.IsolationLevel);
 
-		var sb = new StringBuilder();
-		sb.AppendLine("Releasing savepoint:");
-		sb.AppendLine($"Isolation Level: {transaction.IsolationLevel}");
-		log.Debug(sb.ToString());
-	}
-	public static void TransactionReleasedSavepoint(IFbLogger log, FbTransaction transaction)
-	{
-		if (!log.IsEnabled(FbLogLevel.Debug))
-			return;
+	public static void TransactionRolledBack(ILogger log, FbTransaction transaction) =>
+		log.LogDebug("Rolled back transaction: {isolationLevel}", transaction.IsolationLevel);
 
-		var sb = new StringBuilder();
-		sb.AppendLine("Released savepoint:");
-		sb.AppendLine($"Isolation Level: {transaction.IsolationLevel}");
-		log.Debug(sb.ToString());
-	}
-	public static void TransactionRollingBackSavepoint(IFbLogger log, FbTransaction transaction)
-	{
-		if (!log.IsEnabled(FbLogLevel.Debug))
-			return;
+	public static void TransactionSaving(ILogger log, FbTransaction transaction) =>
+		log.LogDebug("Creating savepoint: {isolationLevel}", transaction.IsolationLevel);
 
-		var sb = new StringBuilder();
-		sb.AppendLine("Rolling back savepoint:");
-		sb.AppendLine($"Isolation Level: {transaction.IsolationLevel}");
-		log.Debug(sb.ToString());
-	}
-	public static void TransactionRolledBackSavepoint(IFbLogger log, FbTransaction transaction)
-	{
-		if (!log.IsEnabled(FbLogLevel.Debug))
-			return;
+	public static void TransactionSaved(ILogger log, FbTransaction transaction) =>
+		log.LogDebug("Created savepoint: {isolationLevel}", transaction.IsolationLevel);
 
-		var sb = new StringBuilder();
-		sb.AppendLine("Rolled back savepoint:");
-		sb.AppendLine($"Isolation Level: {transaction.IsolationLevel}");
-		log.Debug(sb.ToString());
-	}
-	public static void TransactionCommittingRetaining(IFbLogger log, FbTransaction transaction)
-	{
-		if (!log.IsEnabled(FbLogLevel.Debug))
-			return;
+	public static void TransactionReleasingSavepoint(ILogger log, FbTransaction transaction) =>
+		log.LogDebug("Releasing savepoint: {isolationLevel}", transaction.IsolationLevel);
 
-		var sb = new StringBuilder();
-		sb.AppendLine("Committing (retaining) transaction:");
-		sb.AppendLine($"Isolation Level: {transaction.IsolationLevel}");
-		log.Debug(sb.ToString());
-	}
-	public static void TransactionCommittedRetaining(IFbLogger log, FbTransaction transaction)
-	{
-		if (!log.IsEnabled(FbLogLevel.Debug))
-			return;
+	public static void TransactionReleasedSavepoint(ILogger log, FbTransaction transaction) =>
+		log.LogDebug("Released savepoint: {isolationLevel}", transaction.IsolationLevel);
 
-		var sb = new StringBuilder();
-		sb.AppendLine("Committed (retaining) transaction:");
-		sb.AppendLine($"Isolation Level: {transaction.IsolationLevel}");
-		log.Debug(sb.ToString());
-	}
-	public static void TransactionRollingBackRetaining(IFbLogger log, FbTransaction transaction)
-	{
-		if (!log.IsEnabled(FbLogLevel.Debug))
-			return;
+	public static void TransactionRollingBackSavepoint(ILogger log, FbTransaction transaction) =>
+		log.LogDebug("Rolling back savepoint: {isolationLevel}", transaction.IsolationLevel);
 
-		var sb = new StringBuilder();
-		sb.AppendLine("Rolling back (retaining) transaction:");
-		sb.AppendLine($"Isolation Level: {transaction.IsolationLevel}");
-		log.Debug(sb.ToString());
-	}
-	public static void TransactionRolledBackRetaining(IFbLogger log, FbTransaction transaction)
-	{
-		if (!log.IsEnabled(FbLogLevel.Debug))
-			return;
+	public static void TransactionRolledBackSavepoint(ILogger log, FbTransaction transaction) =>
+		log.LogDebug("Rolled back savepoint: {isolationLevel}", transaction.IsolationLevel);
 
-		var sb = new StringBuilder();
-		sb.AppendLine("Rolled back (retaining) transaction:");
-		sb.AppendLine($"Isolation Level: {transaction.IsolationLevel}");
-		log.Debug(sb.ToString());
-	}
+	public static void TransactionCommittingRetaining(ILogger log, FbTransaction transaction) =>
+		log.LogDebug("Committing (retaining) transaction: {isolationLevel}", transaction.IsolationLevel);
 
-	static bool IsNullParameterValue(object value) => value == DBNull.Value || value == null;
+	public static void TransactionCommittedRetaining(ILogger log, FbTransaction transaction) =>
+		log.LogDebug("Committed (retaining) transaction: {isolationLevel}", transaction.IsolationLevel);
+
+	public static void TransactionRollingBackRetaining(ILogger log, FbTransaction transaction) =>
+		log.LogDebug("Rolling back (retaining) transaction: {isolationLevel}", transaction.IsolationLevel);
+
+	public static void TransactionRolledBackRetaining(ILogger log, FbTransaction transaction) =>
+		log.LogDebug("Rolled back (retaining) transaction: {isolationLevel}", transaction.IsolationLevel);
+
+	private static object NormalizeDbNull(object value) =>
+		value == DBNull.Value || value == null
+			? null
+			: value;
+
+	private static Dictionary<string, object> FbParameterCollectionToDictionary(FbParameterCollection parameters) =>
+		parameters
+			.Cast<DbParameter>()
+			.ToList()
+			.ToDictionary(
+				p => p.ParameterName,
+				p => NormalizeDbNull(p.Value)
+			);
 }