… src/FirebirdSql.Data.FirebirdClient/Logging/ConsoleLoggingProvider.cs
… src/FirebirdSql.Data.FirebirdClient/Logging/FbLogLevel.cs
… src/FirebirdSql.Data.FirebirdClient/Logging/IFbLogger.cs
… src/FirebirdSql.Data.FirebirdClient/Logging/IFbLoggingProvider.cs
… src/FirebirdSql.Data.FirebirdClient/Logging/NullLoggingProvider.cs