The code has been rewritten from the ground up, though I've kept some names and structure, at least for this initial version.

- `IFbLogger` -> `Microsoft.Extensions.Logging.ILogger<T>`
- `FbLogManager` -> Probably better to rename it to `FBLoggingConfiguration`?
- `LogMessages` -> Converted to structured logging format. 

There are some warnings about libraries not tested with `.net5`. What about to support `.net6` and up with compiler directives?

In `.net6` log messages can be [generated in compile-time](https://learn.microsoft.com/en-us/dotnet/core/extensions/logger-message-generator).